rename: user_show.js を jscroll.js に変更
ファイル内の jscroll の設定処理をタイムラインページ・ユーザーページで共有するため、user_show.js のファイル名だと分かりづらい為
